% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_tagging_processing_functions.R
\name{aggregate_detections_butte}
\alias{aggregate_detections_butte}
\title{Aggregate Detections for Butte Region}
\usage{
aggregate_detections_butte(
  detections,
  receiver_metadata,
  create_detection_history = FALSE
)
}
\arguments{
\item{detections}{A dataframe containing detections information.}

\item{receiver_metadata}{A dataframe containing receiver metadata.}
}
\value{
A list containing:
\itemize{
\item A dataframe of detections with replaced receiver locations.
\item A dataframe of aggregated receiver metadata.
}
}
\description{
Replace receiver locations in the detections dataframe with aggregated locations for the Butte region.
}
\details{
This function replaces receiver locations in the detections dataframe with aggregated locations for the Butte region according to the specified replacement dictionary. It aggregates receiver locations into one and calculates the mean River Kilometer (RKM), Latitude, Longitude, and Region for each aggregated location. The function also updates the receiver metadata accordingly.
}
\examples{

# Aggregate detections for Butte region
aggregated_data <- aggregate_detections_butte(detections_data, receiver_metadata)

}
