% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_tagging_processing_functions.R
\name{aggregate_detections_sacramento}
\alias{aggregate_detections_sacramento}
\title{Aggregate Detections Data for Sacramento River Sites}
\usage{
aggregate_detections_sacramento(
  detections,
  receiver_metadata,
  create_detection_history = FALSE
)
}
\arguments{
\item{detections}{A dataframe containing detections information.}

\item{receiver_metadata}{A table containing metadata for general release sites. Created using \code{get_reciever_sites_metadata()}.}
}
\value{
A dataframe of detections with aggregated receiver sites and updated metadata.
}
\description{
Replaces specific receiver locations in the detections dataframe with aggregated sites, updating their River Kilometer (RKM), Latitude (Lat), Longitude (Lon), and Region accordingly.
}
\details{
This function replaces specified receiver locations in the detections dataframe with aggregated sites according to the provided replace dictionary. It updates the River Kilometer (RKM), Latitude (Lat), Longitude (Lon), and Region for the aggregated sites. The function creates a new dataframe containing the aggregated receiver sites' metadata, reflecting the aggregation done.
}
