% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{btspasx_special_priors_data}
\alias{btspasx_special_priors_data}
\title{Special Priors for BT-SPAS-X Abundance Model}
\format{
A dataframe with 71 rows and 4 columns
\itemize{
\item \code{site}: site
\item \code{run_year}: run year
\item \code{week}: week
\item \code{special_prior}: log normalized special prior abundance cap
}
}
\usage{
btspasx_special_priors_data
}
\description{
A dataframe containing special priors to fine tune model
}
\keyword{datasets}
