% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carcass_estimates}
\alias{carcass_estimates}
\title{Carcass Survey CJS Estimates}
\format{
\itemize{
\item \code{col 1}:col 1
}
}
\usage{
carcass_estimates
}
\description{
Carcass Estimates, Cormack-Jolly-Seber model used to generate
carcass estimates from raw carcass data - TODO add source to show / explain specific sites we have this for
}
\keyword{datasets}
