% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Convert Celsius to Fahrenheit}
\usage{
celsius_to_fahrenheit(celsius)
}
\arguments{
\item{celsius}{A numeric value representing the temperature in Celsius.}
}
\value{
A numeric value representing the temperature in Fahrenheit.
}
\description{
This function converts temperature from Celsius to Fahrenheit.
}
\examples{
celsius_to_fahrenheit(0) # Should return 32
celsius_to_fahrenheit(100) # Should return 212

}
