% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chosen_site_years_to_model}
\alias{chosen_site_years_to_model}
\title{Years to Include in Modeling}
\format{
Dataframe with 186 rows and 8 columns
\itemize{
\item \code{monitoring_year} : year
\item \code{stream}: Stream RST is located on
\item \code{site}: Site RST is located on
\item \code{site_group}: Site group
\item \code{min_date} : Minimum date of sampling season to include
\item \code{min_week} : Minimum week of sampling season to include
\item \code{max_date} : Maximum date of sampling season to include
\item \code{max_week} : Maximum week of sampling season to include
}
}
\source{
Expert opinion. See \code{vignette("years_to_include_analysis", package = "SRJPEdata")} for more details.
}
\usage{
chosen_site_years_to_model
}
\description{
Datasets containing the chosen sites and years of monitoring data to include in model.
See \code{vignette("years_to_include_analysis", package = "SRJPEdata")} for more details
}
\keyword{datasets}
