% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_yearling_ruleset}
\alias{daily_yearling_ruleset}
\title{Daily Yearling Rulesets}
\format{
Dataframe...
\itemize{
\item \code{monitoring_year} : year
}
}
\usage{
daily_yearling_ruleset
}
\description{
Datasets containing the daily yearling rulesets.
See \code{vignette("yearling_ruleset", package = "SRJPEdata")} for more details.
}
\keyword{datasets}
