% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detection_history_feather_butte}
\alias{detection_history_feather_butte}
\title{Travel Time Model Data - Feather River and Butte Creek}
\format{
\itemize{
\item \code{fish_id}: Unique identifier associated with each released fish
\item \code{TTfR1}: Time in days from detection at the release point to the Sacramento River.
\item \code{TTfR2}: Time in days from detection at the release point to the Endpoint.
\item \code{stream}: Stream associated with the study IDs, either \code{feather} or \code{butte}.
}
@source Central Valley Enhanced Acoustic Tagging Project.
This data package queries tagging data from many studies using the ERDDAP data server.
See \code{data-raw/pull_acoustic_tagging_data.R} and \code{data-raw/create_detection_history.R} for more details.
}
\usage{
detection_history_feather_butte
}
\description{
Acoustic tagging data used to model travel time as part of the survival model for the SR JPE
}
\keyword{datasets}
