% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{environmental_data}
\alias{environmental_data}
\title{Environmental Gage Data}
\format{
\itemize{
\item \code{year}: Year associated with environmental measure
\item \code{week}: Week associated with environmental measure
\item \code{statistic}: Summary static used to summarize flow and temperature measures into a single daily reccord (min, mean, max)
\item \code{value}: Flow or temperature measurements
\item \code{stream}: Stream environmental data is collected on
\item \code{gage_agency}: Agency providing flow or temperature data, most data pulled from CDEC or USGS
\item \code{gage_number}: Unique identifier of gage used to query flow and temperature data
\item \code{parameter}: Parameter measured, includes "flow" and "temperature"
\item \code{site_group}: Site group, used to separate traps within the same stream that have unique environmental conditions.
}
\strong{Flow gages:}
\itemize{
\strong Deer creek
\itemize{
\item USGS - 11383500
\item CDEC - DCV
}
\strong Sacramento River
\itemize{
\item USGS - 11383500
\item USGS - 11377100
}
\strong Battle Creek
\itemize{
\item USGS - 11376550
\item USFWS - UBC
}
\strong Butte Creek
\itemize{
\item USGS - 11376550
\item USGS - BCK
}
\strong Clear Creek
\itemize{
\item USFWS - UCC
\item USGS - 11372000
\item USFWS - LCC
}
\strong Mill Creek
\itemize{
\item USGS - 11381500
\item CDEC - MLM
}
\strong Feather River
\itemize{
\item USGS - 11407000
\item CDEC - TFB
\item CDEC - GRL
\item CDEC - FSB
\item CDEC - FRA
\item interpolated
}
\strong Yuba River
\itemize{
\item USGS - 11421000
\item CDEC - YR7
\item interpolated
}
}
\strong{Temperature gages:}
\itemize{
\strong Deer creek
\itemize{
\item CDEC - DCV
}
\strong Sacramento River
\itemize{
\item USGS - 11383500
}
\strong Battle Creek
\itemize{
\item USFWS - UBC
}
\strong Clear Creek
\itemize{
\item USFWS - UCC
\item USFWS - LCC
}
\strong Mill Creek
\itemize{
\item CDEC - MLM
}
\strong Feather River
\itemize{
\item CDEC - GRL
\item CDEC - FRA
\item interpolated
}
\strong Yuba River
\itemize{
\item CDEC - YR7
\item interpolated
}
}
@source USGS/CDEC/FWS. See \code{data-raw/pull_environmental_data.R} for more details.
}
\usage{
environmental_data
}
\description{
Environmental gage data for each tributary in the SR JPE.
Data includes flow and temperature measurements from gages located on tributaries and the Sacramento River mainstem.
Data were compiled from CDEC, USGS, and USFWS sources and are summarized by week and statistic type.
Interpolation was performed to fill in data gaps where temperature data were missing or the time series was incomplete.
For the Feather River and Yuba River, temperature values were interpolated using regression models based on nearby gages and historical data, allowing for consistent estimation across incomplete records.
}
\keyword{datasets}
