% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{escapement_estimates_all_runs}
\alias{escapement_estimates_all_runs}
\title{Escapement estimates all runns}
\format{
\itemize{
\item \code{run_year}: Year that salmon returned to the tributary.
\item \code{stream}: Stream that salmon returned to. This has been filtered to SR JPE tributaries. See GrandTab for a full list of locations.
\item \code{estimates}: Escapement estimate. Note that these values are generated through a variety of methods.
\item \code{run}: Run of Chinook salmon. Currently only fall are included. See \code{annual_adult} for the best spring run estimates
\item \code{species}: Chinook salmon
\item \code{data_type}: Escapement estimates from GrandTab.
}
\usage{
escapement_estimates_all_runs
}
\description{
This table makes escapement data from GrandTab (https://wildlife.ca.gov/Conservation/Fishes/Chinook-Salmon/Anadromous-Assessment) available for use. Currently only fall run are included.
}
\keyword{datasets}
