% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(fahrenheit)
}
\arguments{
\item{fahrenheit}{A numeric value representing the temperature in Fahrenheit.}
}
\value{
A numeric value representing the temperature in Celsius.
}
\description{
This function converts temperature from Fahrenheit to Celsius.
}
\examples{
fahrenheit_to_celsius(32) # Should return 0
fahrenheit_to_celsius(212) # Should return 100

}
