% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_data}
\alias{flow_data}
\title{Environmental Gage Data - Flow}
\format{
\itemize{
\item \code{year}: Year associated with environmental measure
\item \code{week}: Week associated with environmental measure
\item \code{statistic}: Summary static used to summarize flow measures into a single daily reccord (min, mean, max)
\item \code{value}: Flow measurements
\item \code{stream}: Stream environmental data is collected on
\item \code{gage_agency}: Agency providing flow data, most data pulled from CDEC or USGS
\item \code{gage_number}: Unique identifier of gage used to query flow data
\item \code{parameter}: Parameter measured, "flow"
\item \code{site_group}: Site group, used to separate traps within the same stream that have unique environmental conditions.
}
}
\source{
USGS/CDEC/FWS. See \code{data-raw/pull_data_scripts/pull_flow_data.R} for more details.
}
\usage{
flow_data
}
\description{
Flow gage data for each tributary in the SR JPE.
Data includes flow measurements from gages located on tributaries and the Sacramento River mainstem.
Data were compiled from CDEC, USGS, and USFWS sources and are summarized by week and statistic type.
The temporal coverage of flow data ranges from 1930 to 2025.
}
\details{
Gage Information
\strong{Deer creek}
\itemize{
\itemize{Flow gages - Temporal coverage: 1986 - 2025
\item USGS - 11383500
\item CDEC - DCV
}
}
\strong{Sacramento River}
\itemize{
\itemize{Flow gages - Temporal coverage: 1994 - 2025
\item USGS - 11383500
\item USGS - 11377100
}
}
\strong{Battle Creek}
\itemize{
\itemize{Flow gages - Temporal coverage: 1995 - 2025
\item USGS - 11376550
\item USFWS - UBC
}
}
\strong{Butte Creek}
\itemize{
\itemize{Flow gages - Temporal coverage: 1930 - 2025
\item USGS - 11376550
\item USGS - BCK
}
}
\strong{Clear Creek}
\itemize{
\itemize{Flow gages - Temporal coverage: 1995 - 2025
\item USFWS - UCC
\item USGS - 11372000
\item USFWS - LCC
}
}
\strong{Mill Creek}
\itemize{
\itemize{Flow gages - Temporal coverage: 1995 - 2025
\item USGS - 11381500
\item CDEC - MLM
}
}
\strong{Feather River}
\itemize{
\itemize{Flow gages - Temporal coverage: 1969 - 2025
\item USGS - 11407000
\item CDEC - TFB
\item CDEC - GRL
\item CDEC - FSB
\item CDEC - FRA
\item interpolated
}
}
\strong{Yuba River}
\itemize{
\itemize{Flow gages - Temporal coverage: 1999 - 2025
\item USGS - 11421000
\item CDEC - YR7
\item interpolated
}
}
}
\keyword{datasets}
