% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forecast_covariates}
\alias{forecast_covariates}
\title{Forecast covariates}
\format{
\itemize{
\item \code{name}: Name of the covariate
\item \code{year}: Calendar year. Not all variables are summarized by calendar year and may be NA.
\item \code{water_year}: Water year. Not all variables are summarized by water year and may be NA.
\item \code{stream}: Stream
\item \code{value}: Numeric covariate value
\item \code{month}: Month. Not all variables are available at the monthly level.
\item \code{text_value}: Text value for categorical variables.
}
\usage{
forecast_covariates
}
\description{
Table of potential forecast covariates
}
\keyword{datasets}
