% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_tagging_processing_functions.R
\name{get_receiver_sites_metadata}
\alias{get_receiver_sites_metadata}
\title{Get Receiver Sites Metadata for Given Detections Dataframe}
\usage{
get_receiver_sites_metadata(all_detections)
}
\arguments{
\item{all_detections}{A dataframe containing detections information.}
}
\value{
A dataframe of receiver sites along with River Kilometer (RKM), Latitude (Lat), Longitude (Lon), and receiver_region.
}
\description{
Get a list of all receiver sites and metadata for a given a detections dataframe.
Detections data can be pulled directly from ERDAP using \verb{library(rerddap).} See \code{?pull_detections_data()} for more information.
}
\details{
This function processes the provided detections dataframe to extract unique receiver sites' metadata, including their River Kilometer (RKM), Latitude (Lat), Longitude (Lon), and Region. It calculates mean RKM, Lat, and Lon for each receiver site to handle potential discrepancies in the detections dataframe.
}
\examples{
# Load the detections dataframe
data <- read.csv("detections.csv")

# Get receiver sites metadata
receiver_metadata <- get_receiver_sites_metadata(data)

}
