% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_tagging_processing_functions.R
\name{make_fish_encounter_history}
\alias{make_fish_encounter_history}
\title{Make Fish Encounter History}
\usage{
make_fish_encounter_history(
  detections,
  aggregated_reciever_metadata,
  released_fish_table
)
}
\arguments{
\item{detections}{A dataframe containing detections information.}

\item{aggregated_receiver_metadata}{A dataframe containing aggregated receiver metadata.}
}
\value{
An encounter history dataframe. A matrix of every fish tagged for a given studyID
at every given receiver site (that is in reach.meta.aggregate) and whether it was present 1 or absent 0 in the detection dataframe.
}
\description{
Create an encounter history dataframe based on detections and aggregated receiver metadata.
}
\details{
This function creates an encounter history dataframe, which is a matrix of every fish tagged for a given studyID at every given receiver site (that is in reach.meta.aggregate), indicating whether the fish was present (1) or absent (0) in the detection dataframe. It utilizes the earliest detection time for each fish at each receiver site to determine presence or absence.
}
\examples{
# Create encounter history dataframe
encounter_history <- make_fish_encounter_history(detections_data, aggregated_receiver_metadata)

}
