% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{observed_adult_input}
\alias{observed_adult_input}
\title{Combined Adult Survey and Passage Counts}
\format{
\itemize{
\item \code{year}: year data were collected
\item \code{stream}: stream data were collected on
\item \code{data_type}: type of adult data (either \code{upstream_estimate},
\code{redd_count}, \code{holding_count}, or \code{carcass_estimate})
\item \code{count}: aggregated yearly count of fish
}
}
\source{
Adult data packages are currently being added to EDI.
Data in this table was pulled from holding, redd, carcass, and passage datasets.
See \code{?SRJPEdata::holding}, \code{?SRJPEdata::redd}, \code{?SRJPEdata::carcass_estimates},
and \code{?SRJPEdata::passage_estimates} for more details.
}
\usage{
observed_adult_input
}
\description{
Contains all adult holding and redd survey counts, CJS estimates,
and interpolated upstream passage estimates aggregated by year. Includes all
available years for all streams for each data type. This can be used as input for
the Passage to Spawner model.
}
\keyword{datasets}
