% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_tagging_processing_functions.R
\name{pull_detections_data_from_ERDDAP}
\alias{pull_detections_data_from_ERDDAP}
\title{Pull Detections Data Based on Study IDs}
\usage{
pull_detections_data_from_ERDDAP(study_id)
}
\arguments{
\item{study_id}{A character study IDs. For all ERDDAP valid study IDs see \code{?pull_study_ids_from_ERDDAP()}}
}
\value{
A dataframe containing detections data from the ERDDAP table.
}
\description{
Pulls detections data from a specified ERDDAP table based on a list of study IDs.
}
\details{
This function retrieves detections data from the 'FED_JSATS_detects' ERDDAP table based on the provided list of study IDs. It constructs a query to filter detections by the specified study IDs and fetches the corresponding data from the ERDDAP server.
}
\examples{
# Define a list of study IDs
study_id <- "study_id_1"

# Pull detections data based on study IDs
detections_data <- pull_detections_data(study_id)

}
