% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_tagging_processing_functions.R
\name{pull_fish_data_from_ERDDAP}
\alias{pull_fish_data_from_ERDDAP}
\title{Pull Fish Data from ERDDAP}
\usage{
pull_fish_data_from_ERDDAP()
}
\value{
A dataframe containing tagged fish data from the ERDDAP table.
}
\description{
Pulls tagged fish data from a specified ERDDAP table.
}
\details{
This function retrieves tagged fish data from the 'FED_JSATS_taggedfish' ERDDAP table. It fetches the data from the specified ERDDAP server URL and performs necessary data transformations, including converting date columns to POSIXct format and extracting fish ID prefixes.
}
\examples{
# Pull fish data from ERDDAP
fish_data <- pull_fish_data_from_ERDDAP()

}
