% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_tagging_processing_functions.R
\name{pull_reciever_data_from_ERDDAP}
\alias{pull_reciever_data_from_ERDDAP}
\title{Pull Receiver Data from ERDDAP}
\usage{
pull_reciever_data_from_ERDDAP()
}
\value{
A dataframe containing receiver data from the ERDDAP table.
}
\description{
Pulls receiver data from the FED_JSATS_receivers table.
}
\details{
This function retrieves receiver data from the 'FED_JSATS_receivers' ERDDAP table. It fetches the data from the specified ERDDAP server URL.
}
\examples{
# Pull receiver data from ERDDAP
receiver_data <- pull_receiver_data_from_ERDDAP()

}
