% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_tagging_processing_functions.R
\name{pull_study_ids_from_ERDDAP}
\alias{pull_study_ids_from_ERDDAP}
\title{Pull Study IDs from ERDDAP}
\usage{
pull_study_ids_from_ERDDAP()
}
\value{
A character vector containing unique study IDs.
}
\description{
Pulls unique study IDs from a specified ERDDAP table.
}
\details{
This function retrieves unique study IDs from the 'FED_JSATS_detects' ERDDAP table. It fetches the data from the specified ERDDAP server URL and extracts the 'study_id' field, ensuring uniqueness of the study IDs.
}
\examples{
# Pull study IDs from ERDDAP
study_ids <- pull_study_ids_from_ERDDAP()

}
