% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redd}
\alias{redd}
\title{Redd Survey Data}
\format{
\itemize{
\item \code{date}: year data were collected
\item \code{latitude}: latitude of redd location
\item \code{longitude}: longitude of redd location
\item \code{reach}: survey reach where data were collected
\item \code{river_mile}: river mile of redd location
\item \code{redd_id}: unique identifier of redd
\item \code{fish_guarding}: whether or not fish were observed guarding the redd
\item \code{redd_measured}: whether or not the redd was measured
\item \code{redd_width}: widdth of redd (m)
\item \code{redd_length}: length of redd (m)
\item \code{velocity}: measured stream velocity (ft/s)
\item \code{age}: age of redd
\item \code{age_index}: number of times a unique redd was aged If 0, the redd was surveyed but not aged.
\item \code{redd_count}: number of redds surveyed
\item \code{stream}: stream data were collected on
\item \code{year}: year data were collected
\item \code{survey_method}: surveying method
\item \code{run}: run of fish associated with redd
\item \code{species}: species of fish associated with redd
\item \code{depth_m}: depth of water (m)
\item \code{starting_elevation_ft}: starting elevation (ft)
\item \code{num_of_fish_on_redd}: number of fish observed on redd
\item \code{redd_substrate_class}: size class of substrate determined by millimeter size
\item \code{tail_substrate_class}: size class of tail substrate determined by millimeter size
\item \code{pre_redd_substrate_class}: size class of substrate pre-redd determined by millimeter size
}
}
\source{
Redd data were provided by monitoring programs.
These data are currently being uploaded to EDI. In the meantime you can find data exploration scripts
for each tributary \href{https://github.com/SRJPE/JPE-datasets/tree/main/data-raw/qc-markdowns/adult-holding-redd-and-carcass-surveys}{here} and
combined redd data \href{https://github.com/SRJPE/JPE-datasets/blob/main/data-raw/standard-format-data-prep/standard_adult_redd_data.Rmd}{here}.
}
\usage{
redd
}
\description{
Raw redd survey monitoring data. Redd survey data is available
for Battle Creek, Clear Creek, Feather River, Mill Creek, Yuba River.
TODO link to vignette on age_index when available
}
\keyword{datasets}
