% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{release_fish}
\alias{release_fish}
\title{RST Efficiency Trial Individual Fish Release Data}
\format{
Dataframe with \strong{TODO EMPTY IN DATABASE CURRENTLY} and 6 columns
\itemize{
\item \code{release_id}: year
\item \code{stream}: stream where efficiency trial conducted
\item \code{site}: site where efficiency trial conducted
\item \code{subsite}: subsite where efficiency trial conducted
\item \code{site_group}: site group where efficiency trial conducted
\item \code{fork_length}: fork length of released fish
}
}
\usage{
release_fish
}
\description{
Dataset containing individual fish releases for efficiency monitoring data for all SR JPE tribuatary and mainstem monitoring programs
}
\keyword{datasets}
