% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rst_catch}
\alias{rst_catch}
\title{RST Catch Data}
\format{
A dataframe with 511,961 rows and 14 columns
\itemize{
\item \code{date}: date
\item \code{stream}: Stream RST is located on
\item \code{site}: Site RST is located on
\item \code{subsite}: Specific trap site
\item \code{site_group}: Site group
\item \code{count}: Number of fish caught
\item \code{run}: Run of catch
\item \code{life_stage}: Life Stage of catch
\item \code{adipose_clipped}: Boolean value describing if adipose is clipped on catch
\item \code{dead}: Mortality status of catch
\item \code{fork_length}: Fork length measure of catch in mm
\item \code{weight}: Weight of catch in grams
\item \code{actual_count}: Boolean Value describing if count is actual value or interpolated
\item \code{species}: Species of catch
}
}
\usage{
rst_catch
}
\description{
Dataset containing rst catch monitoring data for all SR JPE tribuatary and mainstem monitoring programs
}
\keyword{datasets}
