% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rst_cwt_recaptures}
\alias{rst_cwt_recaptures}
\title{RST Recaptures}
\format{
\itemize{
\item \code{date}: Date of capture
\item \code{forklength}: Fork length of catch in millimeters
\item \code{weight}: Weight of catch in grams
\item \code{tag_code}: Coded wire tag code
\item \code{stream}: River (Sacramento River)
\item \code{site}: Site of recapture (Knights Landing RST)
\item \code{week_index}: Week index (aligns with btspasx model where week for is week 1, etc.)
\item \code{release_group_id}: Id variable created to group releases that have the same characteristics together. This can be used to join with the hatchery_release table.
}
\usage{
rst_cwt_recaptures
}
\description{
Table of recaptures from Knights Landing RST
}
\keyword{datasets}
