% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rst_model_years}
\alias{rst_model_years}
\title{RST Modeling Years}
\format{
\itemize{
\item \code{stream}: Stream
\item \code{site}: RST monitoring site
\item \code{run_year}: Run year (week >= 45 is included in the following calendar year; e.g. Dec 1995 would be included in run_year 1996)
\item \code{number_weeks}: Number of weeks sampled in the run_year, for reference
\item \code{exclusion_type}: Provides the exclusion ruleset applied
\item \code{apply_to}: Differentiates between years that can be excluded broadly ("all runs") or those specific to spring run ("spring")
\item \code{exclude}: Boolean value to describe if a year should be excluded or not
}
}
\usage{
rst_model_years
}
\description{
Dataframe to compile all years, streams, & sites of RST data and indicate if a year should be used for modeling.
}
\keyword{datasets}
