% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rst_trap}
\alias{rst_trap}
\title{RST Trap Data}
\format{
\itemize{
\item \code{trap_start_date}: Date that trap was visited. This field is included because in some cases a visit will not have a start/stop time (e.g. if trap was cleaned/serviced or just checked)
\item \code{visit_type}:
\item \code{trap_stop_date}: Date that trap was started prior to being sampled. trap_start_date and trap_stop_date are meant to describe the sampling period.
\item \code{trap_start_time}: Time that trap was visited.
\item \code{trap_stop_time}:  Time that trap was sampled.
\item \code{stream}: Which stream the RST is located on
\item \code{site}: Site name
\item \code{subsite}: Name of trap
\item \code{site_group}: Site group, used to seperate traps within the same stream that have unique environmental conditions.
\item \code{trap_functioning}: Describes how well trap is functioning. Categories: trap functioning normally, trap stopped functioning, trap not in service, trap functioning but not normally. If trap_functioning not used, "not recorded" is entered.
\item \code{in_half_cone_configuration}: Boolean to describe if trap fished in half cone configuration. If not recorded, then assumed to be FALSE.
\item \code{fish_processed}: Describes if fish were processed. Categories: processed fish; no fish caught; no catch data, fish released; no catch data, fish left in live box. If fish_processed not used, "not recorded" is entered.
\item \code{rpm_start}: Revolutions per minute at the start of the trap visit before trap is cleaned. If not recorded, then NA.
\item \code{rpm_end}:  Revolutions per minute at the end of the trap visit after trap is cleaned. If not recorded, then NA.
\item \code{total_revolutions}: Total revolutions during sample period. If not recorded, then NA.
\item \code{debris_volume}: Volume of debris emptied from trap (gallons). If not recorded, then NA.
\item \code{discharge}: Flow at trap in cfs
\item \code{water_velocity}: Velocity at trap in feet per second
\item \code{water_temp}: Temperature at trap in degrees C
\item \code{include}: Boolean to describe if sample should be included in data for analysis. If false then data or visit is determined to be of poor quality by data steward and should not be included in analysis. If not recorded, then assumed to be TRUE.
\item \code{turbidity}: Turbidity at trap in NTUs
}
}
\source{
Raw datasets and original documentation of data processing and QC process can be found for each stream using the links below.
\itemize{
\item \strong{Battle Creek}: \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.1509.1}{Data on EDI}, \href{https://github.com/SRJPE/JPE-datasets/blob/main/data-raw/qc-markdowns/rst/battle-creek/battle_creek_rst_environmental_qc.md}{Original Data Exploration Script}. Additional data exploration scripts are located in GitHub.
\item \strong{Butte Creek}: \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.1497.1}{Data on EDI}, \href{https://github.com/SRJPE/JPE-datasets/blob/main/data-raw/qc-markdowns/rst/butte-creek/butte-creek-rst-qc-checklist.md}{Original Data Exploration Script}. Additional data exploration scripts are located in GitHub.
\item \strong{Clear Creek}: \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.1509.1}{Data on EDI}, \href{https://github.com/SRJPE/JPE-datasets/blob/main/data-raw/qc-markdowns/rst/clear-creek/clear_creek_rst_environmental.md}{Original Data Exploration Script}. Additional data exploration scripts are located in GitHub.
\item \strong{Deer Creek}: \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.1504.1}{Data on EDI}, \href{https://github.com/SRJPE/JPE-datasets/blob/main/data-raw/qc-markdowns/rst/deer-creek/deer_creek_rst_data_qc.md}{Original Data Exploration Script}. Additional data exploration scripts are located in GitHub.
\item \strong{Feather River}: \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.1239.5}{Data on EDI}, \href{https://github.com/SRJPE/JPE-datasets/blob/main/data-raw/qc-markdowns/rst/feather-river/feather-rst-effort.md}{Original Data Exploration Script}. Additional data exploration scripts are located in GitHub.
\item \strong{Mill Creek}: \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.1504.1}{Data on EDI}, \href{https://github.com/SRJPE/JPE-datasets/blob/main/data-raw/qc-markdowns/rst/mill-creek/mill_creek_rst_qc.md}{Original Data Exploration Script}. Additional data exploration scripts are located in GitHub.
\item \strong{Sacramento River}:
\itemize{
\item{Knights Landing}: \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.1501.1}{Data on EDI}, \href{https://github.com/SRJPE/JPE-datasets/tree/main/data-raw/qc-markdowns/rst/lower-sac/knights_landing/}{Original Data Exploration Script}. Additional data exploration scripts are located in GitHub.
\item{Tisdale}: \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.1499.2}{Data on EDI}, \href{https://github.com/SRJPE/JPE-datasets/tree/main/data-raw/qc-markdowns/rst/lower-sac/tisdale}{Original Data Exploration Script}. Additional data exploration scripts are located in GitHub.
\item{Delta Entry}: \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.1503.1}{Data on EDI}.
}
\item \strong{Yuba River}: \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.1529.2}{Data on EDI}, \href{https://github.com/SRJPE/JPE-datasets/blob/main/data-raw/qc-markdowns/rst/yuba-river/yuba-river-rst-qc-checklist.md}{Original Data Exploration Script}. Additional data exploration scripts are located in GitHub.
}
}
\usage{
rst_trap
}
\description{
Dataset containing rst trap operations monitoring data for all SR JPE tribuatary and mainstem monitoring programs
}
\keyword{datasets}
