% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_lookup}
\alias{site_lookup}
\title{Site Lookup}
\format{
\itemize{
\item \code{stream}: Stream RST is located on
\item \code{site}: Site RST is located on
\item \code{subsite}: Specific trap site
\item \code{site_group}: Site group, used to separate traps within the same stream that have unique environmental conditions.
}
\usage{
site_lookup
}
\description{
Site lookup table for linking subsite, sites, site_groups, and streams in the SR JPE
}
\keyword{datasets}
