% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_recruit_covariates}
\alias{stock_recruit_covariates}
\title{Stock recruit covariates}
\format{
\itemize{
\item \code{year}: Calendar year.
\item \code{stream}: Stream
\item \code{lifestage}: Lifestage that the covariate applies to
\item \code{covariate_type}: Describes the type of covariate (temperature or flow)
\item \code{covariate_structure}: Describes the structure of the covariate.
\item \code{value}: Numeric value
}
\usage{
stock_recruit_covariates
}
\description{
Table of covariates used in stock recruit modeling
}
\keyword{datasets}
