% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survival_model_inputs}
\alias{survival_model_inputs}
\title{Survival Model Data}
\format{
\itemize{
\item \code{fish_id}: Unique identifier associated with each released fish
\item \code{ch}: Capture history of each fish describing detection at receiver location, 0 indicated not detected, 1 indicates detected
\item \code{study_id}: Unique identifier associated with each release group
\item \code{fish_length}: Fork length of fish in mm
\item \code{fish_weight}: Weight of fish in grams
\item \code{fish_type}: Type of fish used in release trial, typically indicated where that fish came from and species of fish
\item \code{fish_release_date}: Date of release
\item \code{release_location}: Location of release
\item \code{year}: Year of release
}
@source Central Valley Enhanced Acoustic Tagging Project.
This data package queries tagging data from many studies using the ERDDAP data server.
See \code{data-raw/pull_acoustic_tagging_data.R} for more details.

\itemize{
\item \code{fish_id}: Unique identifier associated with each released fish
\item \code{ch}: Capture history of each fish describing detection at receiver location, 0 indicated not detected, 1 indicates detected
\item \code{study_id}: Unique identifier associated with each release group
\item \code{fish_length}: Fork length of fish in mm
\item \code{fish_weight}: Weight of fish in grams
\item \code{fish_type}: Type of fish used in release trial, typically indicated where that fish came from and species of fish
\item \code{fish_release_date}: Date of release
\item \code{release_location}: Location of release
\item \code{year}: Year of release
}
@source Central Valley Enhanced Acoustic Tagging Project.
This data package queries tagging data from many studies using the ERDDAP data server.
See \code{data-raw/pull_acoustic_tagging_data.R} for more details.

\itemize{
\item \code{fish_id}: Unique identifier associated with each released fish
\item \code{ch}: Capture history of each fish describing detection at receiver location, 0 indicated not detected, 1 indicates detected
\item \code{study_id}: Unique identifier associated with each release group
\item \code{fish_length}: Fork length of fish in mm
\item \code{fish_weight}: Weight of fish in grams
\item \code{fish_type}: Type of fish used in release trial, typically indicated where that fish came from and species of fish
\item \code{fish_release_date}: Date of release
\item \code{release_location}: Location of release
\item \code{year}: Year of release
}
@source Central Valley Enhanced Acoustic Tagging Project.
This data package queries tagging data from many studies using the ERDDAP data server.
See \code{data-raw/pull_acoustic_tagging_data.R} for more details.
}
\usage{
survival_model_inputs

survival_model_inputs

survival_model_inputs
}
\description{
Acoustic tagging data used to model survival for the SR JPE

Acoustic tagging data used to model survival for the SR JPE

Acoustic tagging data used to model survival for the SR JPE
}
\keyword{datasets}
