% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_data}
\alias{temperature_data}
\title{Environmental Gage Data - Temperature}
\format{
\itemize{
\item \code{year}: Year associated with environmental measure
\item \code{week}: Week associated with environmental measure
\item \code{statistic}: Summary static used to summarize temperature measures into a single daily reccord (min, mean, max)
\item \code{value}: Temperature measurements
\item \code{stream}: Stream environmental data is collected on
\item \code{gage_agency}: Agency providing temperature data, most data pulled from CDEC or USGS
\item \code{gage_number}: Unique identifier of gage used to query temperature data
\item \code{parameter}: Parameter measured, "temperature"
\item \code{site_group}: Site group, used to separate traps within the same stream that have unique environmental conditions.
}
}
\source{
USGS/CDEC/FWS. See \code{data-raw/pull_data_scripts/pull_temperature_data.R} for more details.
}
\usage{
temperature_data
}
\description{
Temperature gage data for each tributary in the SR JPE.
Data includes temperature measurements from gages located on tributaries and the Sacramento River mainstem.
Data were compiled from CDEC, USGS, and USFWS sources and are summarized by week and statistic type.
Interpolation was performed to fill in data gaps where temperature data were missing or the time series was incomplete.
For the Feather River and Yuba River, temperature values were interpolated using regression models based on nearby gages and historical data, allowing for consistent estimation across incomplete records.
The temporal coverage of temperature data ranges from 1994 to 2025.
}
\details{
Gage Information
\strong Deer creek
\itemize{
\itemize{Temperature gages - Temporal coverage: 1998 - 2025
\item CDEC - DCV
}
}
\strong Sacramento River
\itemize{
\itemize{Temperature gages - Temporal coverage: 1994 - 2023
\item USGS - 11383500
}
}
\strong Battle Creek
\itemize{
\itemize{Temperature gages  - Temporal coverage: 2003 - 2021
\item USFWS - UBC
}
}
\strong Clear Creek
\itemize{
\itemize{Temperature gages - Temporal coverage: 2001 - 2021
\item USFWS - UCC
\item USFWS - LCC
}
}
\strong Mill Creek
\itemize{
\itemize{Temperature gages - Temporal coverage: 1998 - 2025
\item CDEC - MLM
}
}
\strong Feather River
\itemize{
\itemize{Temperature gages - Temporal coverage: 1999 - 2025
\item CDEC - GRL
\item CDEC - FRA
\item interpolated
}
}
\strong Yuba River
\itemize{
\itemize{Temperature gages - Temporal coverage: 1998 - 2025
\item CDEC - YR7
\item interpolated
}
}
}
\keyword{datasets}
