% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upstream_passage_estimates}
\alias{upstream_passage_estimates}
\title{Adult Upstream passage monitoring estimates}
\format{
\itemize{
\item \code{year}: year
}
}
\usage{
upstream_passage_estimates
}
\description{
Adult upstream passage estimates, Generalized Additive Model (GAM)
used to generate passage estimates from raw passage data - TODO add source to show
this / explain different status for different sites
}
\keyword{datasets}
