% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weekly_juvenile_abundance_catch_data}
\alias{weekly_juvenile_abundance_catch_data}
\title{Weekly BT-SPAS-X Catch Data}
\format{
A tibble with 10,890 rows and 14 columns
\itemize{
\item \code{year}: year
\item \code{week} : week
\item \code{stream}: rst stream
\item \code{site}: rst site
\item \code{count}: weekly count of number of
\item \code{mean_fork_length}: Mean fork length of released efficiency trial fish
\item \code{hours_fished}: Hours fished on a stream
\item \code{flow_cfs}: Weekly mean gage flow in cubic feet per second (cfs)
\item \code{log_flow_cfs}: Weekly mean gage flow, log-transformed
\item \code{life_stage}: Life stage of juvenile fish
\item \code{average_stream_hours_fished}: Average hours fished on a given stream, used to standardize
\item \code{standardized_flow}: Flow standardized across years within sites
\item \code{log_standardized_flow}: Log-transformed flow standardized across years within sites
\item \code{run_year}: Run year
\item \code{catch_standardized_by_hours_fished}: Catch standardized by hours fished in a given week
\item \code{lgN_prior}: log normalized special prior abundance cap
}
}
\source{
Prepared using rotary screw trap catch data (see \code{?SRJPEdata::rst_catch} for more information on raw data sources),
flow_data (see \code{?SRJPEdata::flow_data}), hours fished information
(see \verb{?SRJPEdata::weekly_hours_fished for more information}).
}
\usage{
weekly_juvenile_abundance_catch_data
}
\description{
Weekly Rotary Screw Trap (RST) catch summaries, weekly RST effort effort, and weekly standardized flow,
for spring run tributaries modeled in BT-SPAS-X.
}
\keyword{datasets}
