% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weekly_juvenile_abundance_model_data}
\alias{weekly_juvenile_abundance_model_data}
\title{Weekly BT-SPAS-X Model Data}
\format{
A tibble with 9,457 rows and 17 columns
\itemize{
\item \code{year}: year
\week \code{week} : week
\item \code{stream}: rst stream
\item \code{site}: rst site
\item \code{count}:
\item \code{mean_fork_length}:
\item \code{number_released}:
\item \code{number_recaptured}:
\item \code{effort}:
\item \code{flow_cfs}:
\item \code{life_stage}:
\item \code{average_stream_effort}:
\item \code{standardized_flow}:
\item \code{run_year}:
\item \code{catch_standardized_by_effort}:
\item \code{standardized_efficiency}:
\item \code{lgN_prior}:
}
}
\usage{
weekly_juvenile_abundance_model_data
}
\description{
Weekly Rotary Screw Trap (RST) catch summaries, weekly RST effort effort, and weekly standardized flow,
for spring run tributaries modeled in BT-SPAS-X. TODO - consider removing count, effort, and flow_cfs since we have somewhat duplicative columns
}
\keyword{datasets}
