% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{years_to_exclude_rst_data}
\alias{years_to_exclude_rst_data}
\title{Years to exclude - RST data}
\format{
\itemize{
\item \code{stream}: Stream
\item \code{site}: RST monitoring site
\item \code{run_year}: Run year (week >= 45 is included in the following calendar year; e.g. Dec 1995 would be included in run_year 1996)
\item \code{number_weeks}: Number of weeks sampled in the run_year, for reference
\item \code{exclusion_type}: Provides the exclusion ruleset applied
\item \code{apply_to}: Differentiates between years that can be excluded broadly ("all runs") or those specific to spring run ("spring")
}
\usage{
years_to_exclude_rst_data
}
\description{
Years for RST data that were identified as incomplete and not used in modeling
}
\keyword{datasets}
