% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{years_to_include_rst_data}
\alias{years_to_include_rst_data}
\title{Years to include - RST data, applies to all runs}
\format{
\itemize{
\item \code{stream}: Stream
\item \code{site}: RST monitoring site
\item \code{run_year}: Run year (week >= 45 is included in the following calendar year; e.g. Dec 1995 would be included in run_year 1996)
}
\usage{
years_to_include_rst_data
}
\description{
Years for RST data that can be included in modeling (inverse of years to exclude where apply_to == "all runs")
}
\keyword{datasets}
