% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_yearling_ruleset}
\alias{daily_yearling_ruleset}
\title{Daily Yearling Rulesets}
\format{
Dataframe with 2,565 rows and 4 columns (stream, month, day, cutoff)
\itemize{
\item \code{stream}: Stream RST is located on
\item \code{month}: Month to apply yearling ruleset to
\item \code{day}: Day to apply yearling ruleset to
\item \code{cutoff}: Fork length cutoff value in mm
}
}
\source{
Expert opinion. See \code{vignette("yearling_ruleset", package = "SRJPEdata")} for more details.
}
\usage{
daily_yearling_ruleset
}
\description{
Datasets containing the daily yearling rulesets.
See \code{vignette("yearling_ruleset", package = "SRJPEdata")} for more details.
}
\keyword{datasets}
