% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efficiency_summary}
\alias{efficiency_summary}
\title{RST Efficiency Summary Data}
\format{
\itemize{
\item \code{year}: year
}
}
\usage{
efficiency_summary
}
\description{
Dataset containing summarized efficiency monitoring data for all SR JPE tribuatary and mainstem monitoring programs
}
\keyword{datasets}
