% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{environmental_data}
\alias{environmental_data}
\title{Environmental Gage Data}
\format{
\itemize{
\item \code{year}: Year associated with environmental measure
\item \code{week}: Week associated with environmental measure
\item \code{statistic}: Summary static used to summarize flow and temperature measures into a single daily reccord (min, mean, max)
\item \code{value}: Flow or temperature measurements
\item \code{stream}: Stream environmental data is collected on
\item \code{gage_agency}: Agency providing flow or temperature data, most data pulled from CDEC or USGS
\item \code{gage_number}: Unique identifier of gage used to query flow and temperature data
\item \code{parameter}: Parameter measured, includes "flow" and "temperature"
\item \code{site_group}: Site group, used to separate traps within the same stream that have unique environmental conditions.
}
@source USGS/CDEC/FWS. See \code{data-raw/pull_environmental_data.R} for more details.
}
\usage{
environmental_data
}
\description{
Environmental gage data for each tributary in the SR JPE.
}
\keyword{datasets}
