% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rst_trap}
\alias{rst_trap}
\title{RST Trap Data}
\format{
\itemize{
\item \code{year}: year
}
}
\usage{
rst_trap
}
\description{
Dataset containing rst trap operations monitoring data for all SR JPE tribuatary and mainstem monitoring programs
}
\keyword{datasets}
