% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upstream_passage_estimates}
\alias{upstream_passage_estimates}
\title{Adult Upstream passage monitoring estimates}
\format{
\itemize{
\item \code{year}: year data were collected
\item \code{stream}: stream data were collected on
\item \code{passage_estimate}: yearly aggregated interpolated upstream passage estimate
\item \code{run}: run of fish (if assigned)
\item \code{adipose_clipped}: whether or not the adipose fin was clipped
}
}
\source{
Adult upstream passage estimates were provided by monitoring programs.
These data are currently being uploaded to EDI. In the meantime you can find data exploration scripts
for each tributary \href{https://github.com/SRJPE/JPE-datasets/tree/main/data-raw/qc-markdowns/adult-upstream-passage-monitoring}{here} and
combined passage estimates \href{https://github.com/SRJPE/JPE-datasets/blob/main/data-raw/standard-format-data-prep/adult_passage_estimate_standard_format.Rmd}{here}.
}
\usage{
upstream_passage_estimates
}
\description{
Adult upstream passage estimates, Generalized Additive Model (GAM)
used to generate passage estimates from raw passage data. Upstream passage estimates
are available for: battle creek, clear creek, deer creek, mill creek, yuba river, butte creek.
}
\keyword{datasets}
