% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upstream_passage_estimates}
\alias{upstream_passage_estimates}
\alias{annual_adult}
\title{Annual adult estimates}
\format{
\itemize{
\item \code{year}: year data were collected (calendar year)
\item \code{stream}: stream data were collected on
\item \code{count}: annual estimate
\item \code{data_type}: type of data (redd, carcass estimate, holding, passage estimate)
\item \code{lower_bound_estimate}: if confidence intervals are available, value for the lower bound estimate
\item \code{upper_bound_estimate}: if confidence intervals are available, value for the upper bound estimate
\item \code{confidence_level}: if confidence intervals are available, level of confidence provided
}
}
\source{
Annual adult estimates were provided by monitoring programs.
These data are currently being uploaded to EDI.
}
\usage{
annual_adult
}
\description{
Annual adult estimates by data type provided by monitoring programs.
}
\keyword{datasets}
