% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weekly_juvenile_abundance_efficiency_data}
\alias{weekly_juvenile_abundance_efficiency_data}
\title{Weekly BT-SPAS-X Model Efficency Data}
\format{
A tibble with 651 rows and 8 columns
\itemize{
\item \code{year}: year
\week \code{week} : week
\item \code{stream}: rst stream
\item \code{site}: rst site
\item \code{number_released}: Number of efficiency trial fish released
\item \code{number_recaptured}: Number of efficiency trial fish recaptured
\item \code{standardized_efficiency_flow}: Flow standardized across years and sites
\item \code{run_year}: Run year
}
}
\source{
Prepared using rotary screw trap catch data (see \code{?SRJPEdata::rst_catch} for more information on raw data sources),
environmental_data (see \code{?SRJPEdata::environmental_data}), hours fished information
(see \verb{?SRJPEdata::weekly_hours_fished for more information}), and btspasx special priors (see \code{?SRJPEdata::btspasx_special_priors_data} for more information).
}
\usage{
weekly_juvenile_abundance_efficiency_data
}
\description{
Weekly Rotary Screw Trap (RST) efficiency data and weekly standardized flow,
for spring run tributaries modeled in BT-SPAS-X.
}
\keyword{datasets}
