% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weekly_juvenile_abundance_model_data}
\alias{weekly_juvenile_abundance_model_data}
\title{Weekly BT-SPAS-X Model Data}
\format{
A tibble with 9,457 rows and 17 columns
\itemize{
\item \code{year}: year
\week \code{week} : week
\item \code{stream}: rst stream
\item \code{site}: rst site
\item \code{count}: weekly count of number of
\item \code{mean_fork_length}: Mean fork length of released efficiency trial fish
\item \code{number_released}: Number of efficiency trial fish released
\item \code{number_recaptured}: Number of efficiency trial fish recaptured
\item \code{hours_fished}: Hours fished on a stream
\item \code{flow_cfs}: Weekly mean gage flow in cubic feet per second (cfs)
\item \code{life_stage}: Life stage of juvenile fish
\item \code{average_stream_effort}: Average hours fished on a given stream, used to standardize
\item \code{standardized_flow}: Flow standardized across years and sites
\item \code{run_year}: Run year
\item \code{catch_standardized_by_hours_fished}: Catch standardized by hours fished in a given week
\item \code{standardized_efficiency}: Efficiency standardized across years
\item \code{lgN_prior}: log normalized special prior abundance cap, for more info search \code{?SRJPEdata::btspasx_special_priors_data}
}
}
\source{
Prepared using rotary screw trap catch data (see \code{?SRJPEdata::rst_catch} for more information on raw data sources),
environmental_data (see \code{?SRJPEdata::environmental_data}), hours fished information
(see \verb{?SRJPEdata::weekly_hours_fished for more information}), and btspasx special priors (see \code{?SRJPEdata::btspasx_special_priors_data} for more information).
}
\usage{
weekly_juvenile_abundance_model_data
}
\description{
Weekly Rotary Screw Trap (RST) catch summaries, weekly RST effort effort, and weekly standardized flow,
for spring run tributaries modeled in BT-SPAS-X.
}
\keyword{datasets}
