% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{years_to_exclude_rst_data}
\alias{years_to_exclude_rst_data}
\alias{years_to_exclude_adult}
\title{Years to exclude - RST}
\format{
\itemize{
\item \code{year}: Calendar year
\item \code{stream}: Stream
\item \code{data_type}: Type of adult data
\item \code{reason_for_exclusion}: Provides reason for exclusion
}
\usage{
years_to_exclude_adult
}
\description{
Years for RST monitoring locations that were identified as incomplete and not used in modeling
}
\keyword{datasets}
